database refactoring
