+++ src/main/resources/db/seed.sql
@@ -1,3 +1,3 @@
-INSERT INTO Preferences (Code, Value) VALUES
+INSERT OR IGNORE INTO Preferences (Code, Value) VALUES
 ('Language', 'eng.json'),
 ('Theme', 'light');
